Create merge_barcodes.sh
A batch file to merge fail and pass barcode files. 

Note: It is necessary to merge barcode files from multiple sequencing runs or platforms to obtain a larger amount of data for a single sample. This can be useful to improve the accuracy and completeness of the final assembly, or to allow for more in-depth analysis of the sample.